requirements.txt
cycler==0.10.0
joblib==1.0.0
kiwisolver==1.3.1
matplotlib==3.3.4
numpy==1.20.0
Pillow==8.1.0
pyparsing==2.4.7
python-dateutil==2.8.1
scikit-learn==0.24.1
scipy==1.6.0
six==1.15.0
sklearn==0.0
threadpoolctl==2.1.0
